# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface TranslateInput {
  text: string;
  target_language: string;
}

export interface TranslateResult {
  translated: string;
}

export interface ReplyTranslateInput {
  original_message: string;
  your_reply: string;
  target_language: string;
}

export interface ReplyTranslateResult {
  translated_reply: string;
}