# lib/api-zod/src/generated/types/replyTranslateInput.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ReplyTranslateInput {
  original_message: string;
  your_reply: string;
  target_language: string;
}