# lib/api-zod/src/generated/types/translateInput.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface TranslateInput {
  text: string;
  target_language: string;
}